refactor(notification): group SerializedStyles by generality
