+++ src/components/Notification/subcomponents/CompactNotification/CompactNotification.style.ts
@@ -3,7 +3,7 @@ import { rem, transparentize } from 'polished';
 import { NotificationTypes } from '../../Notification';
 import { css, SerializedStyles } from '@emotion/core';
 
-export const typeToThemePalette = (theme: Theme, type: NotificationTypes) =>
+export const typeToThemePalette = (theme: Theme, type: NotificationTypes): string =>
   theme.utils.getColor(type, 400, 'normal');
 
 export const notificationsContainer = (withFilling: boolean, type: NotificationTypes) => (
@@ -34,10 +34,6 @@ export const actionsContainer = () => (theme: Theme): SerializedStyles => css`
   padding-right: ${theme.spacing.md};
 `;
 
-export const iconContainer = () => (theme: Theme): SerializedStyles => css`
-  padding-right: ${theme.spacing.sm};
-`;
-
 export const headContainer = () => (theme: Theme): SerializedStyles => css`
   padding-right: ${theme.spacing.xsm};
   font-weight: ${theme.typography.weights.bold};
@@ -46,9 +42,4 @@ export const headContainer = () => (theme: Theme): SerializedStyles => css`
 export const primaryActionContainer = () => (theme: Theme): SerializedStyles => css`
   cursor: pointer;
   color: ${theme.utils.getColor('lightBlue', 400)};
-`;
-
-export const closeActionContainer = () => (theme: Theme): SerializedStyles => css`
-  cursor: pointer;
-  margin-left: ${theme.spacing.lg};
 `;
